DS-414 seperate components and styles to seperate files
