+++ ngx-fudis/projects/ngx-fudis/src/lib/components/tab-navigation/tab-navigation-bar.component.html
@@ -1,0 +1,17 @@
+<div #tabNavigation class="fudis-tab-navigation-bar">
+  <div *ngIf="_leftContentHidden()" class="fudis-tab-navigation-bar__button-container fudis-tab-navigation-bar__button-container--left">
+    <button (click)="_scrollLeft()" aria-hidden="true" tabindex="-1" [class]="'fudis-tab-navigation-bar__button-container__button fudis-tab-navigation-bar__button-container__button--left'">
+      <fudis-icon class="fudis-tab-navigation-bar__button-container__button__icon" rotate="flip-180" icon="chevron"></fudis-icon>
+    </button>
+  </div>
+  <div [class]="'fudis-tab-navigation-bar__wrapper fudis-tab-navigation-bar__wrapper--' + variant">
+    <div #scrollContainer [class]="'fudis-tab-navigation-bar__wrapper__tab-container fudis-tab-navigation-bar__wrapper__tab-container--' + variant">
+      <ng-content></ng-content>
+    </div>
+  </div>
+  <div *ngIf="_rightContentHidden()" class="fudis-tab-navigation-bar__button-container fudis-tab-navigation-bar__button-container--right">
+    <button (click)="_scrollRight()" aria-hidden="true" tabindex="-1" [class]="'fudis-tab-navigation-bar__button-container__button fudis-tab-navigation-bar__button-container__button--right'">
+      <fudis-icon class="fudis-tab-navigation-bar__button-container__button__icon" icon="chevron"></fudis-icon>
+    </button>
+  </div>
+</div>
