--- ngx-fudis/projects/ngx-fudis/src/lib/components/tab-navigation/tab-navigation-tab.component.html
@@ -1,1 +1,0 @@
-<ng-content></ng-content>
