--- ngx-fudis/projects/ngx-fudis/src/lib/components/tab-navigation/tab-navigation.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { TabNavigationComponent } from './tab-navigation.component';
-
-describe('TabNavigationComponent', () => {
-  let component: TabNavigationComponent;
-  let fixture: ComponentFixture<TabNavigationComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [TabNavigationComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(TabNavigationComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
