+++ ngx-fudis/projects/ngx-fudis/src/public-api.ts
@@ -93,11 +93,9 @@ export { SectionContentDirective } from './lib/components/section/section-conten
 export { SelectOptionsDirective } from './lib/components/form/select/common/select-options-directive/select-options.directive';
 export { TooltipDirective } from './lib/directives/tooltip/tooltip.directive';
 export { PopoverDirective } from './lib/directives/popover/popover.directive';
-export {
-  TabNavigationComponent,
-  TabNavigationTabComponent,
-  TabNavigationPanelComponent,
-} from './lib/components/tab-navigation/tab-navigation.component';
+export { TabNavigationTabComponent } from './lib/components/tab-navigation/tab-navigation-tab.component';
+export { TabNavigationBarComponent } from './lib/components/tab-navigation/tab-navigation-bar.component';
+export { TabNavigationPanelComponent } from './lib/components/tab-navigation/tab-navigation-panel.component';
 
 /**
  * SERVICES
